+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.6.0 - 2023-02-09
 
 - Overflow-y is now `auto` instead of `scroll`
-- Fix fullscreen overflow scrolling in script editor dialog
+- Fix: Fullscreen overflow scrolling in script editor dialog
+- Fix: Handle result from `parse` callback in code editor
 
 ## 1.5.1 - 2022-12-14